+++ nRF51/rbc_mesh/include/timeslot_handler.h
@@ -124,6 +124,6 @@ void timeslot_queue_async_event(async_event_t* evt);
 uint32_t timeslot_get_remaining_time(void);
 
 /** @brief returns the timestamp the timeslot is set to end at */
-uint32_t timeslot_get_end_time(void);
+uint64_t timeslot_get_end_time(void);
 
 #endif /* _TIMESLOT_HANDLER_H__ */
