+++ nRF51/rbc_mesh/include/transport_control.h
@@ -51,7 +51,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *   there is anything to process.
 */
 
-void transport_control_timeslot_begin(uint32_t global_timer_value);
+void transport_control_timeslot_begin(uint64_t global_timer_value);
 
 /**
 * @brief Force a check for timed out values
